Add wasm support (#1)
* Add wasm support

* try to fix wasm

* maybe fix

* add cloudflare stuff

* add headers

* fix cloudflare pages deployment maybe

* Fix Cloudflare Pages deployment

Switch from cloudflare/pages-action to direct wrangler CLI usage.
This ensures the deployment is explicitly recognized as Pages
rather than defaulting to Workers deployment.



* Ensure Cloudflare Pages project exists before deployment

- Install wrangler locally to ensure correct version
- Create Pages project if it doesn't exist
- Use npx wrangler for all commands



* Use Wrangler v4 for Cloudflare Pages deployment

The wrangler-action@v3 uses Wrangler v3 which doesn't properly
support static-only Pages deployments. Switching to Wrangler v4
installed directly via npm resolves the "Missing entry-point" error.

References:
- https://github.com/cloudflare/wrangler-action/issues/369



* Allow deployment on both push and pull requests

Enable Cloudflare Pages preview deployments for PRs
and branch pushes to test before merging to main.



* Create Cloudflare Pages project before deployment

Add step to create the Pages project if it doesn't exist.
This fixes the "Project not found" error on first deployment.



---------

+++ pmtiles/pmtiles.hpp
@@ -522,7 +522,7 @@ inline std::tuple<std::string, std::string, int> make_root_leaves(const std::fun
 }
 
 inline void collect_entries(const std::function<std::string(const std::string &, uint8_t)> decompress, std::vector<entry_zxy> &tile_entries, const char *pmtiles_map, const headerv3 &h, uint64_t dir_offset, uint64_t dir_len) {
-	std::string dir_s{pmtiles_map + dir_offset, dir_len};
+	std::string dir_s{pmtiles_map + dir_offset, static_cast<size_t>(dir_len)};
 	std::string decompressed_dir = decompress(dir_s, h.internal_compression);
 
 	auto dir_entries = pmtiles::deserialize_directory(decompressed_dir);
@@ -558,7 +558,7 @@ inline std::pair<uint64_t, uint32_t> get_tile(const std::function<std::string(co
 	uint64_t dir_offset = h.root_dir_offset;
 	uint32_t dir_length = h.root_dir_bytes;
 	for (int depth = 0; depth <= 3; depth++) {
-		std::string dir_s{pmtiles_map + dir_offset, dir_length};
+		std::string dir_s{pmtiles_map + dir_offset, static_cast<size_t>(dir_length)};
 		std::string decompressed_dir = decompress(dir_s, h.internal_compression);
 		auto dir_entries = pmtiles::deserialize_directory(decompressed_dir);
 		auto entry = find_tile(dir_entries, tile_id);
